+++ Gitagram/View/AccountView/Component/BackgroundDesign/BackTypeTwoView.swift
@@ -1,0 +1,14 @@
+//
+//  BackTypeTwoView.swift
+//  Gitagram
+//
+//  Created by 伊藤璃乃 on 2024/06/22.
+//
+
+import SwiftUI
+
+struct BackTypeTwoView: View {
+    var body: some View {
+        Color.BackBlueblackColor
+    }
+}
